+++ CHANGELOG.md
@@ -80,6 +80,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix MTU configuration with wicked. #2248
 - Overlay templates that create multiple files will create the parent
   directories for the files if they don't yet exist
+- Overlay templates that create symbolic links remove an existing link if it
+  already exists in the overlay. #2252
+- Fix overlay creation of symbolic links over an existing file. #2252
 
 ### Dependencies